+++ index.html
@@ -242,7 +242,7 @@
     </section>
   </template>
   <script src="js/main.js"></script>
-  <script src="js/functions.js"></script>
+  <script src="js/utils.js"></script>
 </body>
 
 </html>
